Split Rust code into publishable core crate + PyO3 bindings
Restructure as a Cargo workspace with two crates:
- crates/libibt: pure Rust core with optional `arrow` feature gate,
  publishable to crates.io independently
- crates/libibt-python: PyO3 bindings (publish = false)

Arrow-dependent code (channel.rs, Arrow methods on IbtFile, Arrow error
variant) is feature-gated so the core crate compiles without Arrow.
