+++ crates/libibt/src/session_info.rs
