+++ crates/libibt/src/var_header.rs
